midSquare Method - 3rd Optimization Attempt
Stats: Title Method:
    Unused: 2058
    Total Collisions: 8866
    Time taken: 0.103 seconds
Quote Method:
    Unused: 1349
    Total Collisions: 8157
    Time taken: 0.103 seconds

For this attempt, I used the MidSquare method, where I squared the key given and obtained the 5 middle values, then modded by size to use as my new key to place in the list. Results are above. I chose this approach because I figured it would provide a good randomness to the data, which it did. I could modify it more by having it choose 5 random values from anywhere in the key.