Multiplication Method - 4th optimization attempt
Stats: Title Method:
    Unused: 6814
    Total Collisions: 13622
    Time taken: 0.099 seconds
Quote Method:
    Unused: 8031
    Total Collisions: 14839
    Time taken: 0.099 seconds

For this attempt, I used the Multiplication method, which uses a fixed ratio to multiply the key by. This is then modded to take only the decimal points in the value (e.g., 120.1234567, it only takes the 1234567). Then, it floors the value to get a new key to place. This method worked surprisingly quick, and had many unused buckets, which means it was giving like values. I'll probably remake this one to get more random values.