Mod Method - 1st optimization attempt
Stats: Title Method:
    Unused: 2047
    Total Collisions: 8855
    Time taken: 0.099 seconds
Quote Method:
    Unused: 1345
    Total Collisions: 8153
    Time taken: 0.099 seconds

This method is stupidly simple - it entails getting the key and then modding it by the size of the table. That's it. It was my first method, so I kept it relatively simple. It got pretty good stats, but it had a lot of collisions compared to the unused boxes, so my data was equal only two or three times throughout the list, instead of equal data appearing 10, 20, 30 times.