Fold method - 2nd optimization attempt
Stats: Title Method:
    Unused: 2022
    Total Collisions: 8830
    Time taken: 0.151 seconds
Quote Method:
    Unused: 1336
    Total Collisions: 8144
    Time taken: 0.151 seconds

This was my most complicated method. It splits up each key into slices of 3, then it puts the slices into a list. Then, each number is put back together into a number from 100-999, where they are then added to the next slice of 3. After all that, the number is multiplied by the multiplier and modulo by the size to finally get the newKey. I chose this approach because it was interesting to see what splitting up a list of the key could do to randomize.